requirements.txt
flask
requests
python-dotenv==0.19.1